Removed CodeBlockOld code, and prism-react-renderer dependency
