+++ package.json
@@ -30,7 +30,6 @@
     "moment": "^2.30.1",
     "openai": "^4.40.0",
     "plotly.js": "^2.32.0",
-    "prism-react-renderer": "^2.3.1",
     "react": "^18.3.1",
     "react-bootstrap": "^2.10.2",
     "react-detect-print": "^0.1.2",
