+++ src/layout/CodeBlock.jsx
@@ -1,70 +1,159 @@
-import { useState } from "react";
-import { Highlight, themes } from "prism-react-renderer";
-import colors from "../redesign/style/colors";
-import { CodeBlock } from "../pages/APIDocumentationPage";
+import { useState, useEffect } from "react";
+import { Card, Button, Tooltip } from "antd";
+import { CopyOutlined, DownOutlined, UpOutlined } from "@ant-design/icons";
+import CodeMirror from "@uiw/react-codemirror";
+import { langs } from "@uiw/codemirror-extensions-langs";
+import { espresso } from "thememirror";
 
-export default CodeBlock;
-export function CodeBlockOld({ lines, language }) {
-  const [copyText, setCopyText] = useState("Copy");
-  const code = lines.join("\r\n");
+export default function CodeBlock({
+  data,
+  title,
+  language,
+  maxHeight,
+  showExpand = true,
+}) {
+  const [isExpanded, setIsExpanded] = useState(false);
+  const [isCopied, setIsCopied] = useState(false);
 
-  const handleCopy = () => {
-    navigator.clipboard.writeText(code);
-    setCopyText("Copied!");
-    setTimeout(() => setCopyText("Copy"), 1000);
-  };
+  function handleCopy() {
+    navigator.clipboard.writeText(data);
+    setIsCopied(true);
+  }
+
+  useEffect(() => {
+    if (isCopied) {
+      const timeoutId = setTimeout(() => setIsCopied(false), 1000);
+
+      return () => clearTimeout(timeoutId);
+    }
+  }, [isCopied]);
+
+  const cardTitleComponent = (
+    <div
+      style={{
+        display: "flex",
+        flexDirection: "row",
+        justifyContent: "space-between",
+        alignItems: "center",
+        // Padding is necessary so that button click effect
+        // is visible
+        padding: "0 6px 0 0",
+      }}
+    >
+      <p
+        style={{
+          margin: 0,
+        }}
+      >
+        {language}
+      </p>
+      <div
+        style={{
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "center",
+          gap: "0.5rem",
+        }}
+      >
+        {showExpand && (
+          <Tooltip title={`${isExpanded ? "Close" : "Expand"} the code block`}>
+            <Button
+              type="default"
+              style={{
+                border: 0,
+                fontWeight: 500,
+                boxShadow: "none",
+              }}
+              onClick={() => setIsExpanded((prev) => !prev)}
+            >
+              <div
+                style={{
+                  display: "flex",
+                  flexDirection: "row",
+                  alignItems: "center",
+                  justifyContent: "center",
+                  gap: "0.5rem",
+                }}
+              >
+                {isExpanded ? <UpOutlined /> : <DownOutlined />}
+                <p
+                  style={{
+                    margin: 0,
+                  }}
+                >
+                  {isExpanded ? "Shrink" : "Expand"}
+                </p>
+              </div>
+            </Button>
+          </Tooltip>
+        )}
+        <Tooltip title="Copy the code block">
+          <Button
+            type="default"
+            style={{
+              border: 0,
+              fontWeight: 500,
+              boxShadow: "none",
+              width: "6rem",
+            }}
+            onClick={handleCopy}
+          >
+            <div
+              style={{
+                display: "flex",
+                flexDirection: "row",
+                alignItems: "center",
+                justifyContent: "center",
+                gap: "0.5rem",
+              }}
+            >
+              <CopyOutlined />
+              <p
+                style={{
+                  margin: 0,
+                }}
+              >
+                {isCopied ? "Copied!" : "Copy"}
+              </p>
+            </div>
+          </Button>
+        </Tooltip>
+      </div>
+    </div>
+  );
 
   return (
-    <div style={{ position: "relative" }}>
-      <button
+    <div
+      style={{
+        flex: 1,
+      }}
+    >
+      <div
         style={{
-          position: "absolute",
-          right: 10,
-          top: 10,
-          padding: "5px 10px",
-          borderRadius: "5px",
-          border: "none",
-          backgroundColor: colors.TEAL_ACCENT,
-          color: "white",
-          cursor: "pointer",
-          fontSize: "0.9em",
-          fontFamily: "Arial, sans-serif",
-          transitionDuration: "0.4s",
+          display: "flex",
+          flexDirection: "row",
+          justifyContent: "space-between",
+          alignItems: "center",
         }}
-        onClick={handleCopy}
       >
-        {copyText}
-      </button>
-      <Highlight theme={themes.vsDark} code={code} language={language}>
-        {({ className, style, tokens, getLineProps, getTokenProps }) => (
-          <pre className={className} style={style}>
-            <table>
-              <tbody>
-                {tokens.map((line, i) => (
-                  <tr key={i} {...getLineProps({ line, key: i })}>
-                    <td
-                      style={{
-                        textAlign: "right",
-                        paddingRight: "1em",
-                        paddingLeft: "1.5em",
-                        userSelect: "none",
-                        opacity: "0.5",
-                      }}
-                    >
-                      {i + 1}
-                    </td>
-                    <td>
-                      {line.map((token, key) => (
-                        <span key={key} {...getTokenProps({ token })} />
-                      ))}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </pre>
-        )}
-      </Highlight>
+        <h4>{title}</h4>
+      </div>
+      <Card
+        style={{}}
+        loading={!data}
+        bodyStyle={{
+          padding: 0,
+        }}
+        title={cardTitleComponent}
+      >
+        <CodeMirror
+          value={data}
+          maxHeight={maxHeight ? maxHeight : !isExpanded && "260px"}
+          editable={false}
+          extensions={[language === "json" ? langs.json() : langs.python()]}
+          theme={espresso}
+        />
+      </Card>
     </div>
   );
 }
